+++ lib/services/translations_service/business/flush_cache_drift_use_cases.dart
@@ -1,11 +1,11 @@
 import "package:flutter_riverpod/flutter_riverpod.dart";
 
 import "../../../config/translations_config.dart";
-import "../data/data_source/local/database/translation_db_singleton.dart";
+import "solvro_translator.dart";
 
 extension FlushCacheDriftUseCases on WidgetRef {
   Future<void> flushOldData() async {
-    final db = watch(translationsDatabaseSingletonProvider);
-    await db.deleteOldTranslations(TranslationsConfig.ttl);
+    final translator = watch(solvroTranslatorProvider);
+    await translator.flushOldData(TranslationsConfig.ttl);
   }
 }
