--- lib/services/translations_service/data/data_source/local/database/translation_db_singleton.dart
@@ -1,9 +1,0 @@
-import "package:flutter_riverpod/flutter_riverpod.dart";
-import "package:riverpod_annotation/riverpod_annotation.dart";
-
-import "../translations_data_source.dart";
-
-part "translation_db_singleton.g.dart";
-
-@Riverpod(keepAlive: true)
-TranslationsDataSource translationsDatabaseSingleton(Ref ref) => TranslationsDataSource();
